+++ mcww_web/js/misc/layout.js
@@ -145,7 +145,7 @@ registerPageTabs("options");
 
 onUiLoaded(() => {
     OPTIONS.themeFlags.forEach((mcwwThemeFlag) => {
-        document.body.classList.add(`mcww-theme-flag-${mcwwThemeFlag.toLowerCase()}`);
+        document.body.classList.add(`mcww-theme-flag-${mcwwThemeFlag.toLowerCase().replace(" ", "-")}`);
     });
 });